refactor table interfaces
